backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "FastAPI backend for assignment"
readme = "README.md"
authors = [
    { name = "itsrat17", email = "rachitjaat99@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "bcrypt<5",
    "email-validator>=2.3.0",
    "fastapi>=0.135.1",
    "passlib[bcrypt]>=1.7.4",
    "psycopg[binary]>=3.3.3",
    "pydantic-settings>=2.13.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.48",
    "structlog>=25.5.0",
    "uvicorn[standard]>=0.41.0",
]
[build-system]
requires = ["uv_build>=0.10.8,<0.11.0"]
build-backend = "uv_build"
